+++ tests/testthat/test-trans-compose.R
@@ -24,6 +24,15 @@ test_that("uses breaks from first transformer", {
 test_that("produces informative errors", {
   expect_snapshot(error = TRUE, {
     compose_trans()
-    compose_trans("reverse", "log10")
+    compose_trans("sqrt", "reverse", "log10")
   })
+})
+
+test_that("produces correct domains", {
+  expect_equal(compose_trans("sqrt", "reverse")$domain, c(0, Inf))
+  expect_equal(compose_trans("sqrt", "log")$domain, c(0, Inf))
+  expect_equal(compose_trans("log", "log")$domain, c(1, Inf))
+  expect_equal(compose_trans("reverse", "log")$domain, c(-Inf, 0))
+  expect_equal(compose_trans("reverse", "logit", "log")$domain, c(-1, -0.5))
+  expect_error(compose_trans("sqrt", "reverse", "log")$domain, "invalid domain")
 })